+++ spec/lib/shared_project_specs.rb
@@ -21,19 +21,11 @@ describe 'a project with correct paths', :shared => true do
     @project.paths.config.should == @config
   end
 
-  it 'should set the sources path' do
-    @project.paths.sources.should == @root + 'public/images/sprites/src'
-  end
-
-  it 'should set the sprite path' do
-    @project.paths.sprites.should == @root + 'public/images/sprites'
-  end
-
   it 'should set the SASS output path' do
     @project.paths.sass.should == @root + 'public/stylesheets/sass'
   end
 
   it 'should set the CSS sprite URL' do
-    @project.paths.url.should == '/images/sprites'
+    @project.paths.url.should == '/images/:name.png'
   end
 end